Removed seconds from clock and moved them into presentational
