+++ client/src/components/Presentational/Clock/index.js
@@ -27,7 +27,7 @@ class Clock extends React.Component {
     return (
       <div>
         <h1>Hello, world!</h1>
-        <h2>It is {this.state.date.toLocaleTimeString()}.</h2>
+        <h2>It is {this.state.date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}.</h2>
       </div>
     );
   }
